backend/requirements.txt
fastapi
uvicorn
google-auth-oauthlib
google-api-python-client
pandas
python-dotenv
python-multipart
itsdangerous
requests
